Agentic_workflow/requirements.txt
python-telegram-bot==20.0
openai
llama-index
python-dotenv
supabase
